requirements.txt
# requirements.txt
python-dotenv
requests
google-generativeai
serpapi